# Conflicts:
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/AbstractEpic.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/AbstractUser.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Application.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/ApprovalRule.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Board.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/CreatedChildEpic.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Epic.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/GroupFilter.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/ImpersonationToken.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/IssueFilter.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Iteration.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Job.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Label.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/MergeRequest.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/MergeRequestFilter.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/OauthTokenResponse.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/ProjectAccessToken.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/RelatedEpic.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/Runner.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/SystemHook.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/models/User.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/systemhooks/MergeRequestSystemHookEvent.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/webhook/AbstractPushEvent.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/webhook/BuildCommit.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/webhook/BuildEvent.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/webhook/EventCommit.java
#	gitlab4j-models/src/main/java/org/gitlab4j/api/webhook/PipelineEvent.java

+++ gitlab4j-models/src/main/java/org/gitlab4j/api/models/GroupFilter.java
@@ -21,7 +21,7 @@ public class GroupFilter implements Serializable {
      * List of group IDs to skip in the filter.
      */
     @JsonProperty("skip_groups")
-    private List<Integer> skipGroups;
+    private List<Long> skipGroups;
 
     /**
      * Indicates if all available groups should be included in the filter.
@@ -89,7 +89,7 @@ public class GroupFilter implements Serializable {
      * @param skipGroups List of group IDs to not include in the search
      * @return the reference to this GroupFilter instance
      */
-    public GroupFilter withSkipGroups(List<Integer> skipGroups) {
+    public GroupFilter withSkipGroups(List<Long> skipGroups) {
         this.skipGroups = skipGroups;
         return (this);
     }
